requirements.txt
# Core PDF handling
pymupdf>=1.23.0

# Data validation
pydantic>=2.0.0

# Data manipulation and output
pandas>=2.0.0

# OCR for scanned PDFs
pytesseract>=0.3.10
Pillow>=10.0.0

# Testing
pytest>=7.0.0

# Note: pytesseract requires Tesseract OCR to be installed on your system
# macOS: brew install tesseract
# Ubuntu: sudo apt install tesseract-ocr

